+++ Assets/InfiniteScrollView/Scripts/Runtime/VerticalInfiniteScrollView.cs
@@ -127,16 +127,12 @@ namespace HowTungTung
 
         public override async UniTask Remove(int index, bool withRefresh = true)
         {
-            try
-            {
-                var removeCell = dataList[index];
-                await base.Remove(index, withRefresh);
-                scrollRect.content.anchoredPosition -= new Vector2(0, removeCell.cellSize.y + spacing);
-            }
-            catch (Exception ex)
-            {
-                Debug.LogWarning(ex);
-            }
+            if (index >= dataList.Count)
+                return;
+
+            var removeCell = dataList[index];
+            await base.Remove(index, withRefresh);
+            scrollRect.content.anchoredPosition -= new Vector2(0, removeCell.cellSize.y + spacing);
         }
     }
 }