+++ Assets/InfiniteScrollView/Scripts/CHANGELOG.md
@@ -1,4 +1,8 @@
 ## CHANGELOG
 
+## [1.0.1] - 2023-10-17
+- Fixed cellList count increase bug issue.
+- Optimized InfiniteScrollView.
+
 ## [1.0.0] - 2023-10-16
 - Added InfiniteScrollView with UniTask.